[MNN-36] Explicitly mark that we want to save OTel span events in the auto-instrumentors
Functionality for saving span events when we invoke `add_rag_event_for_span()` was added in https://github.com/lastmile-ai/eval/pull/617. We decided to save this data into the OpenTelemetry span event itself, in order to make the UI consistent and easier to parse information. Setting this argument to true has no functional changes, but it makes it clear that this is an explicit decision and not something that is dependent on how the default implementation works


Pls see comment thread in this PR for context